pyproject.toml
[project]
name = "langChain_tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
]
